+++ crates/wdl-analysis/src/diagnostics.rs
@@ -354,6 +354,13 @@ pub fn type_mismatch(
     )
 }
 
+/// Creates a "non-empty array assignment" diagnostic.
+pub fn non_empty_array_assignment(expected_span: Span, actual_span: Span) -> Diagnostic {
+    Diagnostic::error("cannot assign an empty array to a non-empty array type")
+        .with_label("this is an empty array", actual_span)
+        .with_label("this expects a non-empty array", expected_span)
+}
+
 /// Creates a "call input type mismatch" diagnostic.
 pub fn call_input_type_mismatch(
     types: &Types,