+++ crates/wdl-analysis/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Allow coercion of `Array[T]` to `Array[T]+` unless from an empty array
+  literal ([#213](https://github.com/stjude-rust-labs/wdl/pull/213)).
 * Improved type calculations in function calls and when determining common
   types in certain expressions ([#209](https://github.com/stjude-rust-labs/wdl/pull/209)).
 * Treat a coercion to `T?` for a function argument of type `T` as a preference